Create README.md
My name is Iboro Raymond Akpan. I am a passionate tech enthusiast with a focus on social impact. I leverages my expertise in back-end programming( Python and the FastAPI framework), to empower individuals and organizations working on positive change projects. I believes in the power of technology to actualize visions and i am driven by the desire to bridge the gap between social good and the digital world.